feat: room improvements for faster init for chats (#2294)
* feat: removing unused code, switching to virtual thread, creating async push room utility

* feat: adding single method in room to handle input/output side effects, javadocs for all methods

* fix: room already persists media if media parts created so simplifying operations and not pushing to cloud in each call, moving room asset copying from MessageUtils to RoomUtils

* fix: setting room for insight does not do any cloud operation, fixing filesystem operations to push to cloud async
