+++ src/prerna/engine/impl/model/Room.java
@@ -79,7 +79,6 @@ import prerna.reactor.agent.mcp.MCPUtility.MCPExecution;
 import prerna.sablecc2.PixelRunner;
 import prerna.sablecc2.om.nounmeta.NounMetadata;
 import prerna.theme.PlaygroundThemeUtils;
-import prerna.util.Constants;
 import prerna.util.Utility;
 
 public class Room {
@@ -122,10 +121,31 @@ public class Room {
 	 */
 	private final Map<String, Map<String, Object>> toolLookupByLLMName = new HashMap<>();
 
+	/**
+	 * Creates an empty room instance. Primarily used for serialization frameworks
+	 * and ad-hoc object construction.
+	 */
 	public Room() {
 	}
 
-	// Use this constructor if you want to load from JSON (as from DB)
+	/**
+	 * Creates a room from persisted ROOM-table fields.
+	 *
+	 * @param room_id       room identifier
+	 * @param userId        owning user identifier
+	 * @param roomName      display name
+	 * @param systemMessage ignored legacy parameter retained for constructor
+	 *                      compatibility
+	 * @param projectId     associated project identifier
+	 * @param shareId       share identifier, if shared
+	 * @param isActive      whether the room is active
+	 * @param createdAt     creation timestamp
+	 * @param updatedAt     last update timestamp
+	 * @param messagesJson  persisted message history JSON
+	 * @param pinned        whether the room is pinned
+	 * @param options       room options JSON
+	 * @param modelId       model/engine identifier associated to the room
+	 */
 	public Room(String room_id, String userId, String roomName, String systemMessage, String projectId, String shareId,
 			boolean isActive, Timestamp createdAt, Timestamp updatedAt, String messagesJson, boolean pinned,
 			String options, String modelId) {
@@ -158,27 +178,34 @@ public class Room {
 		}
 	}
 
+	/**
+	 * Re-parses {@link #messagesJson} into the in-memory {@link #messages} list.
+	 */
 	public void parseMessages() {
 		setMessagesFromString(this.messagesJson);
 	}
 
 	/**
-	 * 
-	 * @param msg
-	 * @param systemMessage
-	 * @param modelEngine
-	 * @return
+	 * Sends an input message to the provided model engine using default parent
+	 * resolution and history persistence behavior.
+	 *
+	 * @param msg         input message to send
+	 * @param modelEngine model engine used for inference
+	 * @return assistant response message
 	 */
 	public ResponseMessage ask(InputMessage msg, IModelEngine modelEngine) {
 		return ask(msg, modelEngine, null);
 	}
 
 	/**
-	 * 
-	 * @param msg
-	 * @param modelEngine
-	 * @param parentMessageId
-	 * @return
+	 * Sends an input message to the model engine while explicitly targeting a
+	 * parent message in the conversation tree.
+	 *
+	 * @param msg             input message to send
+	 * @param modelEngine     model engine used for inference
+	 * @param parentMessageId explicit parent message id; when null/blank the latest
+	 *                        message is used
+	 * @return assistant response message
 	 */
 	public ResponseMessage ask(InputMessage msg, IModelEngine modelEngine, String parentMessageId) {
 		Boolean appendToHistory = true;
@@ -186,12 +213,17 @@ public class Room {
 	}
 
 	/**
-	 * 
-	 * @param msg
-	 * @param systemMessage
-	 * @param modelEngine
-	 * @param parentMessageId
-	 * @return
+	 * Core room ask flow.
+	 * <p>
+	 * Builds message payload, applies tool metadata, executes model inference, and
+	 * optionally appends/persists both input and output in room history.
+	 *
+	 * @param msg             input message to send
+	 * @param modelEngine     model engine used for inference
+	 * @param parentMessageId explicit parent message id; when null/blank the latest
+	 *                        message is used
+	 * @param appendToHistory whether to append and persist messages to room history
+	 * @return assistant response message
 	 */
 	public synchronized ResponseMessage ask(InputMessage msg, IModelEngine modelEngine, String parentMessageId,
 			Boolean appendToHistory) {
@@ -201,10 +233,8 @@ public class Room {
 		// if it is full prompt, process that first.
 		if (kwArgMap.containsKey(AbstractModelEngine.FULL_PROMPT)) {
 			AskModelEngineResponse llmResponse = modelEngine.askRoom(msg.getInputPrompt(), this, msg, kwArgMap);
-			ResponseMessage response = ResponseMessage.Builder.fromAskModelEngineResponse(llmResponse).build();
-			response.setRoom(this);
-			MessageUtils.persistMediaPartsToRoomFolder(response, this);
-			return response;
+			applyInputUsageFromModelResponse(msg, llmResponse);
+			return buildAssistantResponseFromModelResponse(llmResponse, modelEngine, msg);
 		}
 
 		// if a specific system message is sent to use, overwrite the existing in the db
@@ -235,19 +265,8 @@ public class Room {
 			kwArgMap.put("message_json", singleMessageJson);
 
 			AskModelEngineResponse llmResponse = modelEngine.askRoom(msg.getInputPrompt(), this, msg, kwArgMap);
-			ResponseMessage response = ResponseMessage.Builder.fromAskModelEngineResponse(llmResponse).build();
-
-			// set transaction id for both pieces
-			msg.setTransactionId(llmResponse.getMessageId());
-			msg.setTokensInMessage(llmResponse.getNumberOfTokensInPrompt());
-			response.setTransactionId(llmResponse.getMessageId());
-
-			response.setModel(modelEngine);
-			response.setRoom(this);
-			response.setParentMessageId(msg.getMessageId());
-			response.setTokensInMessage(llmResponse.getNumberOfTokensInResponse());
-			MessageUtils.persistMediaPartsToRoomFolder(response, this);
-			return response;
+			applyInputUsageFromModelResponse(msg, llmResponse);
+			return buildAssistantResponseFromModelResponse(llmResponse, modelEngine, msg);
 		}
 
 		// if we dont have to keep history. then wipe all previous messages.
@@ -280,20 +299,8 @@ public class Room {
 			kwArgMap.put("message_json", messageJsonString);
 
 			AskModelEngineResponse llmResponse = modelEngine.askRoom(msg.getInputPrompt(), this, msg, kwArgMap);
-			response = ResponseMessage.Builder.fromAskModelEngineResponse(llmResponse).build();
-			response.setMessageId(llmResponse.getMessageId());
-
-			// set transaction id for both pieces
-			msg.setTransactionId(llmResponse.getMessageId());
-			msg.setTokensInMessage(llmResponse.getNumberOfTokensInPrompt());
-			response.setTransactionId(llmResponse.getMessageId());
-
-			// Create the assistant's response message and add to history
-			response.setModel(modelEngine);
-			response.setRoom(this);
-			response.setParentMessageId(msg.getMessageId());
-			response.setTokensInMessage(llmResponse.getNumberOfTokensInResponse());
-			MessageUtils.persistMediaPartsToRoomFolder(response, this);
+			applyInputUsageFromModelResponse(msg, llmResponse);
+			response = buildAssistantResponseFromModelResponse(llmResponse, modelEngine, msg);
 		} catch (Exception e) {
 			classLogger.error("Error running new message in room", e);
 			throw e;
@@ -339,15 +346,26 @@ public class Room {
 	}
 
 	/**
-	 * 
-	 * @param toolCallId
-	 * @param toolName
-	 * @param toolExecutionResponse
-	 * @param toolParameterValues
-	 * @param parentMessageId
-	 * @param modelEngine
-	 * @param insight
-	 * @return
+	 * Adds a tool execution result to the active tool-call context and, when all
+	 * pending tool calls are satisfied, invokes the model for the follow-up
+	 * assistant response.
+	 *
+	 * @param toolCallId            tool call identifier being fulfilled
+	 * @param toolName              tool name used for execution
+	 * @param toolExecutionResponse serialized tool output
+	 * @param toolParameterValues   parameters provided to the tool call
+	 * @param paramValuesMap        model parameter map used when continuing the
+	 *                              assistant turn
+	 * @param parentMessageId       optional parent message id anchoring the branch
+	 * @param modelEngine           model engine used for follow-up inference
+	 * @param insight               insight context used for persistence
+	 * @param toolStatus            execution status for the tool result
+	 * @return model response when all tool calls are fulfilled; otherwise
+	 *         {@code null}
+	 * @throws IllegalStateException    if no compatible tool-call context is
+	 *                                  present
+	 * @throws IllegalArgumentException if {@code toolCallId} does not match the
+	 *                                  current assistant tool-call payload
 	 */
 	public synchronized AskModelEngineResponse addToolExecutionResult(String toolCallId, String toolName,
 			String toolExecutionResponse, Map<String, Object> toolParameterValues, Map<String, Object> paramValuesMap,
@@ -505,13 +523,8 @@ public class Room {
 			ResponseMessage nextAssistant = null;
 			try {
 				llmResponse = modelEngine.askRoom("", this, toolResultsMessage, paramValuesMap);
-				nextAssistant = ResponseMessage.Builder.fromAskModelEngineResponse(llmResponse).build();
-				nextAssistant.setParentMessageId(toolResultsMessage.getMessageId());
-				nextAssistant.setModel(modelEngine);
-				nextAssistant.setTokensInMessage(llmResponse.getNumberOfTokensInResponse());
-
-				// set the input tokens for the input message
-				toolResultsMessage.setTokensInMessage(llmResponse.getNumberOfTokensInPrompt());
+				applyInputUsageFromModelResponse(toolResultsMessage, llmResponse);
+				nextAssistant = buildAssistantResponseFromModelResponse(llmResponse, modelEngine, toolResultsMessage);
 			} catch (Exception e) {
 				// remove the last tool since it failed
 				toolResultsMessage.getParts().removeLast();
@@ -554,9 +567,58 @@ public class Room {
 	}
 
 	/**
-	 * 
-	 * @param parentMessageId
-	 * @return
+	 * Applies prompt-side usage metadata from model output to the originating input
+	 * message.
+	 *
+	 * @param inputMessage input message that triggered the model call
+	 * @param llmResponse  model response containing prompt token count and message
+	 *                     id
+	 */
+	private void applyInputUsageFromModelResponse(InputMessage inputMessage, AskModelEngineResponse llmResponse) {
+		if (inputMessage == null || llmResponse == null) {
+			return;
+		}
+		inputMessage.setTransactionId(llmResponse.getMessageId());
+		inputMessage.setTokensInMessage(llmResponse.getNumberOfTokensInPrompt());
+	}
+
+	/**
+	 * Builds and finalizes an assistant response message from a model response with
+	 * consistent side effects.
+	 * <p>
+	 * Applies response id/transaction, model/room/parent linkage, response token
+	 * count, and persists media parts to the room folder.
+	 *
+	 * @param llmResponse   model response
+	 * @param modelEngine   model engine used for inference
+	 * @param parentMessage parent message to link as this response's parent
+	 * @return finalized assistant response message
+	 */
+	private ResponseMessage buildAssistantResponseFromModelResponse(AskModelEngineResponse llmResponse,
+			IModelEngine modelEngine, AbstractMessage parentMessage) {
+		ResponseMessage response = ResponseMessage.Builder.fromAskModelEngineResponse(llmResponse).build();
+		String llmMessageId = llmResponse.getMessageId();
+		if (llmMessageId != null && !llmMessageId.isEmpty()) {
+			response.setMessageId(llmMessageId);
+			response.setTransactionId(llmMessageId);
+		} else if (response.getTransactionId() == null || response.getTransactionId().isEmpty()) {
+			response.setTransactionId(response.getMessageId());
+		}
+		response.setModel(modelEngine);
+		response.setRoom(this);
+		response.setParentMessageId(parentMessage == null ? null : parentMessage.getMessageId());
+		response.setTokensInMessage(llmResponse.getNumberOfTokensInResponse());
+		RoomUtils.persistMediaPartsToRoomFolder(response, this);
+		return response;
+	}
+
+	/**
+	 * Indicates whether this room can accept a new user input message at the
+	 * current point in conversation state.
+	 *
+	 * @return {@code true} when the last message is a terminal assistant response
+	 *         (text/media), {@code false} when tool-call continuation is required
+	 * @throws IllegalStateException if the room has no messages
 	 */
 	public boolean roomCanAcceptNewInputMessage() {
 		if (messages.isEmpty()) {
@@ -582,9 +644,10 @@ public class Room {
 	}
 
 	/**
-	 * 
-	 * @param params
-	 * @param modelEngine
+	 * Appends room-level tools into the model invocation parameter map.
+	 *
+	 * @param params      mutable model parameter map
+	 * @param modelEngine model engine used to determine max tool name length
 	 */
 	private void appendToolsToParams(Map<String, Object> params, IModelEngine modelEngine) {
 		int maxLength = MCPUtility.getMaxToolNameLength(modelEngine);
@@ -791,6 +854,13 @@ public class Room {
 		return Collections.unmodifiableMap(toolLookupByLLMName);
 	}
 
+	/**
+	 * Checks whether the specified message id belongs to an assistant-authored
+	 * visible output message in this room.
+	 *
+	 * @param messageId message id to validate
+	 * @return {@code true} when a matching assistant output message exists
+	 */
 	public boolean isMessageAuthor(String messageId) {
 		return getMessages().parallelStream()
 				.anyMatch(m -> m.getMessageId().equals(messageId)
@@ -804,9 +874,10 @@ public class Room {
 	 * Returns the effective system prompt by checking options.instructions, then
 	 * workspace.system_prompt, then optionally applying an enterprise-level
 	 * template from the active admin theme.
-	 * 
-	 * @param
-	 * @return String the system prompt or null if none is defined
+	 *
+	 * @return resolved system prompt, or {@code null} when no prompt is configured
+	 * @throws IllegalArgumentException when workspace prompt resolution fails
+	 *                                  access or active-state checks
 	 */
 	public String getEffectiveSystemPrompt() {
 		// 1. Try options.instructions
@@ -862,6 +933,15 @@ public class Room {
 		return expandSystemPromptVariables(merged);
 	}
 
+	/**
+	 * Merges an enterprise-wide system prompt template with the room-level prompt.
+	 *
+	 * @param enterpriseTemplate    enterprise template (may include
+	 *                              {@code {{SYSTEM_PROMPT}}})
+	 * @param effectiveSystemPrompt room/workspace prompt to merge into template
+	 * @return merged prompt string, or {@code null} when no non-blank content
+	 *         remains
+	 */
 	private static String applyEnterpriseSystemPromptTemplate(String enterpriseTemplate, String effectiveSystemPrompt) {
 		enterpriseTemplate = StringUtils.trimToNull(enterpriseTemplate);
 		effectiveSystemPrompt = StringUtils.trimToNull(effectiveSystemPrompt);
@@ -881,6 +961,13 @@ public class Room {
 		return StringUtils.trimToNull(merged);
 	}
 
+	/**
+	 * Expands configured system-prompt variables in the given prompt string.
+	 *
+	 * @param systemPrompt prompt containing optional {@code {{VAR}}} placeholders
+	 * @return prompt with resolved variable substitutions, or {@code null} when
+	 *         input is blank
+	 */
 	private String expandSystemPromptVariables(String systemPrompt) {
 		systemPrompt = StringUtils.trimToNull(systemPrompt);
 		if (systemPrompt == null) {
@@ -914,6 +1001,18 @@ public class Room {
 		return StringUtils.trimToNull(out.toString());
 	}
 
+	/**
+	 * Resolves a configured prompt variable (and optional object path) to its
+	 * string representation.
+	 *
+	 * @param varName        configured variable name
+	 * @param path           optional dotted/indexed path off the root variable
+	 *                       value
+	 * @param configuredVars configured variable map (variable name to pixel
+	 *                       expression)
+	 * @param cache          per-resolution cache to avoid duplicate pixel execution
+	 * @return resolved string value, or {@code null} if unresolved
+	 */
 	private String resolveSystemPromptVariableToString(String varName, String path, Map<String, String> configuredVars,
 			Map<String, Object> cache) {
 		Object rootVal = cache.containsKey(varName) ? cache.get(varName)
@@ -926,6 +1025,14 @@ public class Room {
 		return stringifyPromptValue(val);
 	}
 
+	/**
+	 * Computes a configured system-prompt variable by executing its configured meta
+	 * pixel.
+	 *
+	 * @param varName        configured variable name
+	 * @param configuredVars configured variable map
+	 * @return resolved variable value, or {@code null} if missing/unresolvable
+	 */
 	private Object computeSystemPromptVariable(String varName, Map<String, String> configuredVars) {
 		if (varName == null) {
 			return null;
@@ -941,6 +1048,13 @@ public class Room {
 		return null;
 	}
 
+	/**
+	 * Executes a single-statement META pixel and returns the value from the final
+	 * noun result.
+	 *
+	 * @param pixel raw pixel expression
+	 * @return extracted result value, or {@code null} on validation/exec failure
+	 */
 	private Object runMetaPixelForValue(String pixel) {
 		if (this.insight == null) {
 			return null;
@@ -962,11 +1076,18 @@ public class Room {
 			NounMetadata last = results.get(results.size() - 1);
 			return last == null ? null : last.getValue();
 		} catch (Exception e) {
-			classLogger.debug(Constants.STACKTRACE, e);
+			classLogger.debug("Failed to execute META pixel while resolving system prompt variable. Pixel={}", pixel,
+					e);
 			return null;
 		}
 	}
 
+	/**
+	 * Normalizes and validates that the supplied pixel is a safe single statement.
+	 *
+	 * @param pixel raw pixel expression
+	 * @return normalized pixel ending with {@code ;}, or {@code null} if invalid
+	 */
 	private static String normalizeAndValidateSingleStatementPixel(String pixel) {
 		pixel = StringUtils.trimToNull(pixel);
 		if (pixel == null) {
@@ -985,6 +1106,14 @@ public class Room {
 		return pixel;
 	}
 
+	/**
+	 * Resolves a dotted/indexed path (for example {@code a.b[0].c}) from the
+	 * provided root object.
+	 *
+	 * @param root root value to traverse
+	 * @param path dotted/indexed path
+	 * @return resolved object, or {@code null} when traversal fails
+	 */
 	private static Object resolvePath(Object root, String path) {
 		if (root == null || path == null || path.isBlank()) {
 			return root;
@@ -1028,6 +1157,13 @@ public class Room {
 		return cur;
 	}
 
+	/**
+	 * Resolves a map/JSON-object key on the current traversal value.
+	 *
+	 * @param cur current traversal object
+	 * @param key key name to resolve
+	 * @return resolved value, or {@code null} when key/object is unsupported
+	 */
 	private static Object resolveKey(Object cur, String key) {
 		if (cur == null) {
 			return null;
@@ -1056,6 +1192,13 @@ public class Room {
 		return null;
 	}
 
+	/**
+	 * Resolves a list/array index on the current traversal value.
+	 *
+	 * @param cur current traversal object
+	 * @param idx zero-based index
+	 * @return resolved value, or {@code null} when index/object is unsupported
+	 */
 	private static Object resolveIndex(Object cur, int idx) {
 		if (cur == null || idx < 0) {
 			return null;
@@ -1083,6 +1226,12 @@ public class Room {
 		return null;
 	}
 
+	/**
+	 * Converts a resolved prompt variable value into a prompt-safe string.
+	 *
+	 * @param val value to stringify
+	 * @return stringified value, or {@code null} for null/blank string values
+	 */
 	private static String stringifyPromptValue(Object val) {
 		if (val == null) {
 			return null;
@@ -1102,6 +1251,11 @@ public class Room {
 		return StringUtils.trimToNull(String.valueOf(val));
 	}
 
+	/**
+	 * Reads the enterprise/global system prompt template from the active theme.
+	 *
+	 * @return enterprise prompt template, or {@code null} if unavailable
+	 */
 	private static String getEnterpriseSystemPromptTemplateFromActiveTheme() {
 		try {
 			return PlaygroundThemeUtils.getPlaygroundGlobalSystemPrompt();
@@ -1112,78 +1266,174 @@ public class Room {
 
 	// ---- Getters and Setters ----
 
+	/**
+	 * Returns the room identifier.
+	 *
+	 * @return room id
+	 */
 	public String getId() {
 		return room_id;
 	}
 
+	/**
+	 * Sets the room identifier.
+	 *
+	 * @param id room id
+	 */
 	public void setId(String id) {
 		this.room_id = id;
 	}
 
+	/**
+	 * Returns the owning user id.
+	 *
+	 * @return user id
+	 */
 	public String getUserId() {
 		return userId;
 	}
 
+	/**
+	 * Sets the owning user id.
+	 *
+	 * @param userId user id
+	 */
 	public void setUserId(String userId) {
 		this.userId = userId;
 	}
 
+	/**
+	 * Returns the room display name.
+	 *
+	 * @return room name
+	 */
 	public String getRoomName() {
 		return roomName;
 	}
 
+	/**
+	 * Sets the room display name.
+	 *
+	 * @param roomName room name
+	 */
 	public void setRoomName(String roomName) {
 		this.roomName = roomName;
 	}
 
+	/**
+	 * Returns the room share id.
+	 *
+	 * @return share id
+	 */
 	public String getShareId() {
 		return shareId;
 	}
 
+	/**
+	 * Sets the room share id.
+	 *
+	 * @param shareId share id
+	 */
 	public void setShareId(String shareId) {
 		this.shareId = shareId;
 	}
 
+	/**
+	 * Indicates whether the room is active.
+	 *
+	 * @return {@code true} if active
+	 */
 	public boolean isActive() {
 		return isActive;
 	}
 
+	/**
+	 * Sets whether the room is active.
+	 *
+	 * @param isActive active flag
+	 */
 	public void isActive(boolean isActive) {
 		this.isActive = isActive;
 	}
 
+	/**
+	 * Returns the room creation timestamp.
+	 *
+	 * @return creation timestamp
+	 */
 	public Timestamp getCreatedAt() {
 		return createdAt;
 	}
 
+	/**
+	 * Sets the room creation timestamp.
+	 *
+	 * @param createdAt creation timestamp
+	 */
 	public void setCreatedAt(Timestamp createdAt) {
 		this.createdAt = createdAt;
 	}
 
+	/**
+	 * Returns the room last-updated timestamp.
+	 *
+	 * @return last-updated timestamp
+	 */
 	public Timestamp getUpdatedAt() {
 		return updatedAt;
 	}
 
+	/**
+	 * Sets the room last-updated timestamp.
+	 *
+	 * @param updatedAt last-updated timestamp
+	 */
 	public void setUpdatedAt(Timestamp updatedAt) {
 		this.updatedAt = updatedAt;
 	}
 
+	/**
+	 * Indicates whether the room is pinned.
+	 *
+	 * @return {@code true} if pinned
+	 */
 	public boolean isPinned() {
 		return pinned;
 	}
 
+	/**
+	 * Sets whether the room is pinned.
+	 *
+	 * @param pinned pinned flag
+	 */
 	public void setPinned(boolean pinned) {
 		this.pinned = pinned;
 	}
 
+	/**
+	 * Returns raw room options JSON.
+	 *
+	 * @return options JSON
+	 */
 	public String getOptions() {
 		return options;
 	}
 
+	/**
+	 * Sets raw room options JSON.
+	 *
+	 * @param options options JSON
+	 */
 	public void setOptions(String options) {
 		this.options = options;
 	}
 
+	/**
+	 * Returns parsed room options, lazily deserializing from {@link #options} when
+	 * needed.
+	 *
+	 * @return mutable options map (never {@code null})
+	 */
 	public Map<String, Object> getOptionsMap() {
 		if (optionsMap == null) {
 			if (options != null && !options.trim().isEmpty()) {
@@ -1200,28 +1450,58 @@ public class Room {
 		return optionsMap;
 	}
 
+	/**
+	 * Sets room options from a map and updates the serialized options JSON.
+	 *
+	 * @param map options map
+	 */
 	public void setOptionsMap(Map<String, Object> map) {
 		this.optionsMap = map;
 		this.options = map == null ? null : GSON.toJson(map);
 	}
 
+	/**
+	 * Returns the room model id.
+	 *
+	 * @return model id
+	 */
 	public String getModelId() {
 		return modelId;
 	}
 
+	/**
+	 * Sets the room model id.
+	 *
+	 * @param modelId model id
+	 */
 	public void setModelId(String modelId) {
 		this.modelId = modelId;
 	}
 
+	/**
+	 * Returns the room folder path used for media persistence.
+	 *
+	 * @return room folder path
+	 */
 	public String getRoomFolderPath() {
 		return roomFolderPath;
 	}
 
 	// Core message accessors
+	/**
+	 * Returns the in-memory room message list.
+	 *
+	 * @return mutable message list
+	 */
 	public List<AbstractMessage> getMessages() {
 		return this.messages;
 	}
 
+	/**
+	 * Replaces the in-memory room message list.
+	 *
+	 * @param messagesList message list to set (null clears messages)
+	 */
 	public void setMessages(List<AbstractMessage> messagesList) {
 		this.messages.clear();
 		if (messagesList != null) {
@@ -1229,17 +1509,32 @@ public class Room {
 		}
 	}
 
-	// Serializes the message history to a JSON array for DB storage
+	/**
+	 * Serializes current room messages to DB-safe JSON and updates cached
+	 * {@link #messagesJson}.
+	 *
+	 * @return serialized message JSON
+	 */
 	public String getMessagesAsString() {
-		return MessageUtils.toJsonArray(messages);
+		this.messagesJson = MessageUtils.toJsonArray(messages);
+		return this.messagesJson;
 	}
 
-	// Serializes the message history to a JSON array for python exection
+	/**
+	 * Serializes current room messages to JSON including inline image data for
+	 * model execution payloads.
+	 *
+	 * @return serialized message JSON with image data
+	 */
 	public String getMessagesWithImageDataAsString() {
 		return MessageUtils.toJsonArrayWithImageData(messages);
 	}
 
-	// Deserialize from a JSON string (DB column) and populate the list
+	/**
+	 * Deserializes persisted message JSON into the in-memory message list.
+	 *
+	 * @param messagesJson persisted message JSON from the ROOM table
+	 */
 	public void setMessagesFromString(String messagesJson) {
 		// Pull room folder - Room folder is at BASE_FOLDER/roomid
 		ClusterUtil.pullRoom(this.room_id);
@@ -1252,28 +1547,59 @@ public class Room {
 		this.setMessages(loaded != null ? loaded : new ArrayList<>());
 	}
 
+	/**
+	 * Returns the insight context currently attached to this room.
+	 *
+	 * @return insight context
+	 */
 	public Insight getInsight() {
 		return insight;
 	}
 
+	/**
+	 * Sets the insight context for this room.
+	 *
+	 * @param insight insight context
+	 */
 	public void setInsight(Insight insight) {
 		this.insight = insight;
 	}
 
+	/**
+	 * Returns the cached serialized message JSON value.
+	 *
+	 * @return serialized messages JSON
+	 */
 	public String getMessageJson() {
 		return this.messagesJson;
 	}
 
-	// this should rarely be used. Really only if a Message object was created and
-	// then jsonified
+	/**
+	 * Sets the cached serialized message JSON value directly.
+	 * <p>
+	 * Prefer mutating {@link #messages} and using {@link #getMessagesAsString()}
+	 * unless synchronizing with externally produced JSON.
+	 *
+	 * @param messagesJson serialized messages JSON
+	 */
 	public void setMessagesJson(String messagesJson) {
 		this.messagesJson = messagesJson;
 	}
 
+	/**
+	 * Returns the associated project id.
+	 *
+	 * @return project id
+	 */
 	public String getProjectId() {
 		return projectId;
 	}
 
+	/**
+	 * Sets the associated project id.
+	 *
+	 * @param projectId project id
+	 */
 	public void setProjectId(String projectId) {
 		this.projectId = projectId;
 	}
