+++ src/prerna/reactor/insights/fs/DeleteInsightAssetsReactor.java
@@ -111,7 +111,7 @@ public class DeleteInsightAssetsReactor extends AbstractReactor {
 
 		// push room to cloud storage
 		if (this.insight.getRoomId() != null) {
-			ClusterUtil.pushRoom(this.insight.getRoomId());
+			ClusterUtil.pushRoomAsync(this.insight.getRoomId());
 		}
 
 		NounMetadata retNoun = NounMetadata.getSuccessNounMessage("Success!");
