+++ src/prerna/reactor/insights/fs/NewInsightAssetsDirectoryReactor.java
@@ -85,7 +85,7 @@ public class NewInsightAssetsDirectoryReactor extends AbstractReactor {
 
 		// push room to cloud storage
 		if (this.insight.getRoomId() != null) {
-			ClusterUtil.pushRoom(this.insight.getRoomId());
+			ClusterUtil.pushRoomAsync(this.insight.getRoomId());
 		}
 
 		NounMetadata retNoun = NounMetadata.getSuccessNounMessage("Success!");
