+++ src/prerna/reactor/insights/fs/RenameInsightAssetReactor.java
@@ -95,7 +95,7 @@ public class RenameInsightAssetReactor extends AbstractReactor {
 
 		// push room to cloud storage
 		if (this.insight.getRoomId() != null) {
-			ClusterUtil.pushRoom(this.insight.getRoomId());
+			ClusterUtil.pushRoomAsync(this.insight.getRoomId());
 		}
 
 		NounMetadata retNoun = NounMetadata.getSuccessNounMessage("Success!");
