+++ src/prerna/reactor/model/LLM2Reactor.java
@@ -33,14 +33,12 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import prerna.auth.AccessToken;
 import prerna.auth.User;
 import prerna.auth.utils.SecurityEngineUtils;
 import prerna.engine.api.IModelEngine;
 import prerna.engine.impl.model.Room;
 import prerna.engine.impl.model.RoomUtils;
 import prerna.engine.impl.model.message.InputMessage;
-import prerna.engine.impl.model.message.MessageUtils;
 import prerna.engine.impl.model.message.ResponseMessage;
 import prerna.reactor.AbstractReactor;
 import prerna.sablecc2.om.GenRowStruct;
@@ -53,6 +51,7 @@ import prerna.util.Utility;
 @Deprecated
 public class LLM2Reactor extends AbstractReactor {
 
+	@Deprecated
 	public LLM2Reactor() {
 		this.keysToGet = new String[] { ReactorKeysEnum.ENGINE.getKey(), ReactorKeysEnum.COMMAND.getKey(),
 				ReactorKeysEnum.CONTEXT.getKey(), ReactorKeysEnum.USE_HISTORY.getKey(),
@@ -61,6 +60,7 @@ public class LLM2Reactor extends AbstractReactor {
 		this.keyRequired = new int[] { 1, 1, 0, 0, 0, 0, 0, 0 };
 	}
 
+	@Deprecated
 	@Override
 	public NounMetadata execute() {
 
@@ -69,7 +69,6 @@ public class LLM2Reactor extends AbstractReactor {
 		String engineId = this.keyValue.get(ReactorKeysEnum.ENGINE.getKey());
 		String roomId = this.keyValue.get(ReactorKeysEnum.ROOM_ID.getKey());
 		User user = this.insight.getUser();
-		AccessToken userToken = user.getPrimaryLoginToken();
 
 		if (!SecurityEngineUtils.userCanViewEngine(user, engineId)) {
 			throw new IllegalArgumentException(
@@ -95,12 +94,9 @@ public class LLM2Reactor extends AbstractReactor {
 		List<String> inputImageURLs = getImageURLs();
 
 		Room room = RoomUtils.createRoomIfNotExists(roomId, insight, modelEngine, question);
+		List<String> copiedImages = RoomUtils.copyFilesToRoomFolder(inputImages, room, insight);
 
-		///// MESSAGE CREATION //////////
-
-		List<String> copiedImages = MessageUtils.copyFilesToRoomFolder(inputImages, room, insight);
-		InputMessage msg;
-		msg = InputMessage.builder(room).withSystemPrompt(context).withText(question)
+		InputMessage msg = InputMessage.builder(room).withSystemPrompt(context).withText(question)
 				.withModelType(modelEngine.getModelType()).withParamMap(paramMap).withMediaInputs(copiedImages, room)
 				.withMediaUrls(inputImageURLs).build();
 
@@ -110,6 +106,7 @@ public class LLM2Reactor extends AbstractReactor {
 	}
 
 	// ----------- image/file helpers, paramMap etc. -------------
+	@Deprecated
 	public List<String> getImages() {
 		List<String> inputStrings = new ArrayList<>();
 		GenRowStruct grs = this.store.getGenRowStruct(ReactorKeysEnum.IMAGE.getKey());
@@ -127,6 +124,7 @@ public class LLM2Reactor extends AbstractReactor {
 		return inputStrings;
 	}
 
+	@Deprecated
 	public List<String> getImageURLs() {
 		List<String> inputStrings = new ArrayList<>();
 		GenRowStruct grs = this.store.getGenRowStruct(ReactorKeysEnum.URL.getKey());
@@ -164,11 +162,13 @@ public class LLM2Reactor extends AbstractReactor {
 		return null;
 	}
 
+	@Deprecated
 	@Override
 	public String getReactorDescription() {
 		return "This method is used to run an LLM text-generation call";
 	}
 
+	@Deprecated
 	@Override
 	protected String getDescriptionForKey(String key) {
 		if (key.equals(ReactorKeysEnum.ROOM_ID.getKey())) {
